test: harden the order-search parity probes (review round)
- Trait moved to its own PSR-4-autoloadable file so the HPOS class
  survives a targeted single-file PHPUnit run (it previously depended on
  the sibling file loading first).
- Fixture orders scrub numeric address fields (Order_Address_Scrub_Helpers,
  same as the v1 suites) so the numeric-id LIKE search can never collide
  with a postcode/phone from the shared auto-increment.
- Both classes delete Api::OPTION_ENABLED after parent::tearDown() — the
  flag is written before the transaction starts, so rollback restores
  true and leaked sync routes into later suites.
- HPOS teardown removes only its own __return_true callback instead of
  remove_all_filters.

Tested: both classes 8/8 green, plus the previously-failing targeted
single-file run of the HPOS class (4/4); phpcs clean.
