+++ tests/includes/API/V2/Catalog_Proxy_Order_Search_Tests.php
@@ -1,0 +1,93 @@
+<?php
+/**
+ * Shared V1 order-search parity probes for the v2 catalog proxy.
+ *
+ * @package WCPOS\WooCommercePOS\Tests\API\V2
+ */
+
+namespace WCPOS\WooCommercePOS\Tests\API\V2;
+
+use Automattic\WooCommerce\RestApi\UnitTests\Helpers\OrderHelper;
+use WCPOS\WooCommercePOS\Tests\API\Traits\Order_Address_Scrub_Helpers;
+
+/**
+ * Shared V1 order-search parity probes.
+ */
+trait Catalog_Proxy_Order_Search_Tests {
+	use Order_Address_Scrub_Helpers;
+
+	/**
+	 * Order targeted by the search probes.
+	 *
+	 * @var \WC_Order
+	 */
+	private $target_order;
+
+	/**
+	 * Create orders with distinct billing search fields.
+	 */
+	private function create_order_search_fixtures(): void {
+		$this->target_order = OrderHelper::create_order();
+		$this->target_order->set_billing_first_name( 'AureliaProbe' );
+		$this->target_order->set_billing_last_name( 'QuillonProbe' );
+		$this->target_order->set_billing_email( 'aurelia.order.probe@example.invalid' );
+		// IDs share the global auto-increment: scrub numeric address fields so the
+		// numeric-id LIKE search can never collide with a postcode/phone.
+		$this->scrub_numeric_address_fields( $this->target_order );
+		$this->target_order->save();
+
+		$other_order = OrderHelper::create_order();
+		$other_order->set_billing_first_name( 'BenedictProbe' );
+		$other_order->set_billing_last_name( 'RenshawProbe' );
+		$other_order->set_billing_email( 'benedict.order.probe@example.invalid' );
+		$this->scrub_numeric_address_fields( $other_order );
+		$other_order->save();
+	}
+
+	/**
+	 * Numeric order IDs retain V1 search semantics.
+	 */
+	public function test_order_search_by_id_matches_v1(): void {
+		$this->assert_order_search_finds_target( (string) $this->target_order->get_id() );
+	}
+
+	/**
+	 * Billing first names retain V1 search semantics.
+	 */
+	public function test_order_search_by_billing_first_name_matches_v1(): void {
+		$this->assert_order_search_finds_target( 'AureliaProbe' );
+	}
+
+	/**
+	 * Billing last names retain V1 search semantics.
+	 */
+	public function test_order_search_by_billing_last_name_matches_v1(): void {
+		$this->assert_order_search_finds_target( 'QuillonProbe' );
+	}
+
+	/**
+	 * Partial billing emails retain V1 search semantics.
+	 */
+	public function test_order_search_by_billing_email_matches_v1(): void {
+		$this->assert_order_search_finds_target( 'aurelia.order.probe' );
+	}
+
+	/**
+	 * Dispatch a real V2 request and assert that it finds the target order.
+	 *
+	 * @param string $search Search value.
+	 */
+	private function assert_order_search_finds_target( string $search ): void {
+		$request = $this->wp_rest_get_request( '/wcpos/v2/orders' );
+		$request->set_query_params( array( 'search' => $search ) );
+
+		$response = $this->server->dispatch( $request );
+		$data     = $response->get_data();
+
+		$this->assertEquals( 200, $response->get_status() );
+		$this->assertEquals(
+			array( $this->target_order->get_id() ),
+			wp_list_pluck( $data, 'id' )
+		);
+	}
+}
